fix(frontend): update deps, refactor utb sök for easier debug
