+++ frontend/package.json
@@ -17,63 +17,65 @@
     "jest:coverage": "jest --coverage",
     "test:coverage": "yarn cypress:headless"
   },
+  "dependenciesComments": {
+    "react-hook-form": "v7.55.0 introduces something that creates a submitloop for filter-changes in search"
+  },
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@headlessui/react": "^2.2.0",
     "@hookform/error-message": "^2.0.1",
-    "@hookform/resolvers": "^3.3.4",
-    "@mui/icons-material": "^6.4.7",
-    "@mui/material": "^6.4.7",
-    "@next/bundle-analyzer": "^15.2.1",
-    "@sk-web-gui/core": "3.11.1",
-    "@sk-web-gui/react": "1.4.0",
+    "@hookform/resolvers": "^3.10.0",
+    "@mui/icons-material": "^7.0.1",
+    "@mui/material": "^7.0.1",
+    "@next/bundle-analyzer": "15.2.4",
+    "@sk-web-gui/core": "3.11.2",
+    "@sk-web-gui/react": "1.4.1",
     "@tailwindcss/forms": "^0.5.10",
-    "@tanstack/react-query": "^5.66.0",
-    "axios": "^1.7.9",
-    "bcryptjs": "^2.4.3",
+    "@tanstack/react-query": "^5.71.1",
+    "axios": "^1.8.4",
+    "bcryptjs": "^3.0.2",
     "dayjs": "^1.11.13",
-    "fast-xml-parser": "^4.5.1",
+    "fast-xml-parser": "^4.5.3",
     "html-react-parser": "^5.2.2",
     "lodash": "^4.17.21",
     "lucide-react": "^0.471.1",
-    "next": "15.2.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-hook-form": "^7.54.2",
+    "next": "15.2.4",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-hook-form": "7.54.2",
     "react-i18next": "^15.4.1",
     "react-sticky-el": "^2.1.1",
-    "sanitize-html": "^2.14.0",
+    "sanitize-html": "^2.15.0",
     "sharp": "^0.33.5",
     "swc-plugin-coverage-instrument": "^0.0.26",
     "swiper": "^9.3.2",
     "underscore.string": "^3.3.6",
     "yup": "^1.6.1",
-    "zustand": "^4.5.2"
+    "zustand": "^4.5.6"
   },
   "devDependencies": {
     "@babel/plugin-transform-private-methods": "^7.25.9",
     "@babel/preset-env": "^7.26.9",
     "@babel/preset-react": "^7.26.3",
-    "@babel/preset-typescript": "^7.26.0",
-    "@cypress/code-coverage": "^3.13.12",
+    "@babel/preset-typescript": "^7.27.0",
+    "@cypress/code-coverage": "^3.14.0",
     "@eslint/js": "^8.57.1",
     "@tailwindcss/container-queries": "^0.1.1",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.1",
-    "@types/bcryptjs": "^2.4.6",
     "@types/jest": "^29.5.14",
-    "@types/lodash": "^4.17.15",
-    "@types/node": "22.10.7",
-    "@types/react": "19.0.10",
+    "@types/lodash": "^4.17.16",
+    "@types/node": "22.13.16",
+    "@types/react": "19.0.12",
     "@types/react-dom": "19.0.4",
     "@types/react-sticky-el": "^2.1.0",
-    "@types/sanitize-html": "^2.13.0",
+    "@types/sanitize-html": "^2.15.0",
     "@types/underscore.string": "^0.0.41",
-    "@typescript-eslint/eslint-plugin": "^8.25.0",
-    "autoprefixer": "^10.4.20",
+    "@typescript-eslint/eslint-plugin": "^8.29.0",
+    "autoprefixer": "^10.4.21",
     "babel-jest": "^29.7.0",
     "concurrently": "^9.1.2",
     "cross-env": "^7.0.3",
@@ -81,23 +83,23 @@
     "dotenv": "^16.4.7",
     "envalid": "^8.0.0",
     "eslint": "^8.57.1",
-    "eslint-config-next": "15.2.1",
+    "eslint-config-next": "15.2.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "postcss": "^8.5.2",
+    "postcss": "^8.5.3",
     "postcss-import": "^16.1.0",
     "postcss-loader": "^8.1.1",
     "postcss-nested": "^7.0.2",
     "postcss-nesting": "^13.0.1",
-    "prettier": "^3.4.2",
-    "sass": "^1.84.0",
+    "prettier": "^3.5.3",
+    "sass": "^1.86.1",
     "tailwindcss": "^3.4.17",
-    "typescript": "5.7.3",
-    "typescript-eslint": "^8.25.0",
+    "typescript": "5.8.2",
+    "typescript-eslint": "^8.29.0",
     "xml-js": "^1.6.11"
   },
   "browserslist": [
